API

Tip

All additonal packages used in tests or docs need to be recorded in tests/Project.toml resp. docs/Project.toml, preferably with a [compat] entry.

Documentation

ExampleJuggler.docmodules — Function

     docmodules(example_sources;
                 source_prefix = "https://github.com/j-fu/ExampleJuggler.jl/blob/main/examples")

Generate markdown files for use with documenter from list of Julia code examples. See ExampleLiterate.jl for an example.

ExampleJuggler.@plotmodule — Macro

@plotmodule(modules, kwargs...)

Include module into context of calling module and execute genplots(;kwargs...) if it exists.

ExampleJuggler.@plotmodules — Macro

@plotmodules(modules, kwargs...)

Plot several scripts defining modules via @plotmodule.

Testing

ExampleJuggler.testplutonotebook — Function

testplutonotebook(notebookname; plutoenv = nothing)

Test pluto notebook in a Pluto session.

ExampleJuggler.testplutonotebooks — Function

 testplutonotebooks(notebooks; kwargs...)

Test pluto notebooks via testplutonotebook.

ExampleJuggler.@testmodule — Macro

testmodule(module, kwargs...)

Include script defining a module in the context of the calling module and call the runtests method if it is defined in this module, passing kwargs....

ExampleJuggler.@testmodules — Macro

@testmodules(modules, kwargs...)

Test several scripts defining modules via @testmodule.

ExampleJuggler.@testscript — Macro

testscript(script)

Wrap script (or Pluto notebook seen as script) into a custom module and evaluate it in the context of the calling module. If the script contains a function runtests, call it. It is assumed that the script uses Test.@test to test correctness.

ExampleJuggler.@testscripts — Macro

@testscripts(scripts)

Run scripts in the context of the calling module via @testscript